requirements.txt
# Core ML / Embeddings
sentence-transformers
transformers
torch
numpy

# Document processing
docling
docling-core
pypdf
httpx

# PostgreSQL + Vector search
psycopg[binary,pool]
pgvector

# Data models
pydantic

# CLI
typer
rich
python-dotenv

# Language detection
langdetect

# MCP server
fastmcp

# Dashboard API
fastapi
uvicorn

# Auth
bcrypt
pyjwt
ldap3
